pyproject.toml
[tool.poetry]
name = "firescribe"
version = "0.1.0"
description = "A Python project for migrating, refining, and validating narrative documents."
authors = ["Andrew Kolean <andykolean@gmail.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
openai = "*"
watchdog = "*"
tiktoken = "*"
jsonschema = "*"
pydantic = "*"
pydantic-settings = "*"
pandas = "*"
rich = "*"
sqlalchemy = "*"
numpy = "*"
pyyaml = "^6.0.2"
click = "^8.1.7"
flake8 = "^7.1.1"
flake8-pydantic = "^0.3.1"
flake8-bugbear = "^24.8.19"
flake8-docstrings = "^1.7.0"

[tool.poetry.group.dev.dependencies]
pylance = "^0.18.0"
isort = "^5.13.2"
black = "^24.8.0"
flake8 = "^7.1.1"
mypy = "^1.11.2"
bandit = "^1.7.10"
pytest = "^8.3.3"
python-dotenv = "^1.0.1"
pre-commit = "*"
pytest-cov = "*"
typeguard = "*"
coverage = "*"

[tool.black]
line-length = 79

[tool.isort]
profile = "black"
line_length = 79

[tool.flake8]
max-line-length = 79
ignore = ["E203", "E501", "W503"]
exclude = ["__pycache__", ".venv", "build", "dist"]

[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
disallow_untyped_calls = true
disallow_untyped_defs = true
ignore_missing_imports = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.bandit]
targets = ["src"]
exclude = [".venv", "tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
